Moved color_code function to util.py
